gestion taille données int16

+++ lib/Parser/Helper.hs
@@ -1,12 +1,13 @@
 module Helper where
 
+import Data.Int (Int16)
 import Text.Read (readMaybe)
 
-readMaybeInt :: String -> Either Bool Int
+readMaybeInt :: String -> Either Bool Int16
 readMaybeInt a = maybe (Left False) Right (readMaybe a) >>= applyFilter
   where
-    applyFilter :: Integer -> Either Bool Int
+    applyFilter :: Integer -> Either Bool Int16
     applyFilter b = if filterIntInteger then Right $ fromInteger b else Left True
       where
         filterIntInteger :: Bool
-        filterIntInteger = (b < toInteger (maxBound :: Int)) && (b > toInteger (minBound :: Int))
+        filterIntInteger = (b < toInteger (maxBound :: Int16)) && (b > toInteger (minBound :: Int16))
